Upload local fusionData files to remote Google Drive storage

Convenient and safe wrapper around googledrive functions to upload local file(s) in /fusionData to analogous location in /fusionData directory of "fusionACSdata" Google Drive account.

Usage

uploadFiles(files, ask = TRUE)

Arguments

files

Character. Local file path(s) to be uploaded. Must be within local /fusionData directory.

ask

Logical. Should user be prompted to confirm upload decision in console? Automatically set to FALSE if interactive() == FALSE.

Value

Local files uploaded to Google Drive account. Messages printed to console confirming actions.