Generate /input files needed for fusion

Handles all operations needed to generate /input files from successfully ingested and harmonized donor survey microdata. Also (optionally) uploads resulting local /input data files to correct location in remote storage (via uploadFiles).

NOTE: Argument test_mode is TRUE by default. It must be explicitly set to FALSE to create (or possibly overwrite) production data in local /fusion directory.

Usage

fusionInput(
  donor,
  recipient,
  respondent,
  fuse = NULL,
  force = NULL,
  notes = NULL,
  test_mode = TRUE,
  ncores = getOption("fusionData.cores")
)

Arguments

donor

Character. Donor survey identifier (e.g. "RECS_2015").

recipient

Character. Recipient (ACS) survey identifier (e.g. "ACS_2015").

respondent

Character. Desired respondent level of microdata. Either "household" or "person".

fuse

Character or list. Names of donor variables to be fused to recipient. If fuse is a list, each entry is a character vector possibly indicating multiple variables to fuse as a block. The order of the fuse variables does not matter, since prepXY is used internally to determine a plausible fusion sequence. If NULL (default), an attempt is made to return all donor variables not used in predictor harmonization process.

force

Character. Pre-specified subset of potential predictor variables to "force" as included predictors. These variables are also used within fusionOutput to create validation subsets. We generally select the variables that best reflect the following socioeconomic and geographic concepts: income; race/ethnicity; education; household size; housing tenure; and the highest-resolution location variable for which the donor survey is thought to be representative.

notes

Character. Optional notes supplied by user. These are added to the log file for reference.

test_mode

Logical. If TRUE (default), function uses the local "/fusion_" sub-directory (creating it if necessary). Only when test_mode = FALSE is it possible to overwrite production data in /fusion (no underscore).

ncores

Integer. Number of physical CPU cores used for parallel computation.

Value

Saves resulting /input data files to appropriate local directory and (optionally) to remote Google Drive storage. Also saves a .txt log file alongside data files that records console output from fusionInput().

Details

The following steps: