Generate output files resulting from fusion

Handles all operations needed to "do fusion" using input files generated by a successful call to fusionInput. Trains a fusion model, generates internal validation results, and then simulates multiple implicates for recipient microdata. Also (optionally) uploads resulting local /output data files to correct location in remote storage (via uploadFiles).

Usage

fusionOutput(
  input,
  output = NULL,
  M = NULL,
  note = NULL,
  test_mode = TRUE,
  upload = FALSE,
  ncores = getOption("fusionData.cores"),
  margin = 2,
  ...
)

Arguments

input

Character. Path to directory containing files created by fusionInput.

output

Character. Optional path to directory where output files will be saved. If output = NULL (default), the output directory is automatically constructed from input.

M

Integer. Desired number of fusion implicates. If M = NULL (default) it is internally set to 40 or, if test_mode = TRUE, 2 implicates.

note

Character. Optional note supplied by user. Inserted in the log file for reference.

test_mode

Logical. If test_mode = TRUE (default), the result files are always saved within "/fusion_", faster hyperparameters are used for train, and the internal validation step is skipped.

upload

Logical. Should resulting local /output data files be uploaded to analogous location in remote Google Drive storage? Default is FALSE.

ncores

Integer. Number of physical CPU cores used for parallel computation.

margin

Numeric. Passed to same argument in fuse.

...

Optional, non-default arguments passed to train. For example, fork = TRUE to enable forked parallel processing.

Value

Saves resulting /output data files to appropriate local directory and (optionally) to remote Google Drive storage. Also saves a .txt log file alongside data files that records console output from fusionOutput.

Details

The function checks arguments and determines the file path to the appropriate /output directory (creating it if necessary). The output files are always placed within the appropriate directory hierarchy, based on the donor and recipient information detected in the input file names. In practice, output need only be specified if working in an environment where the output files need to located somewhere different from the input files.

The function executes the following steps:

Load training data inputs. Loads donor training microdata and results of prepXY.

Run fusionModel::train(). Calls train using sensible defaults and hyperparameters. If test_mode = TRUE, the hyperparameters are designed to do a fast/rough-and-ready model training.

Fuse onto training data for internal validation. Fuses multiple implicates to original donor training data using fuse. Results saved to disk. If test_mode = TRUE, this step is skipped.

Run fusionModel::validate(). Passes previous step's results to validate. Results saved to disk. If test_mode = TRUE, this step is skipped.

Fuse onto prediction data. Fuses multiple implicates to supplied input prediction data using fuse. Results saved to disk.

Upload /output files to Google Drive. If upload = TRUE and other requirements are met, uploads output data files to the analogous location in the remote Google Drive storage.

fusionOutput() is finished! Upon completion, a log file named "outputlog.txt" is written to /output for reference.

Examples

# Since 'test_mode = TRUE' by default, this will affect files in local '/fusion_' directory
dir <- fusionInput(donor = "RECS_2015",
                   recipient = "ACS_2015",
                   respondent = "household",
                   fuse = c("btung", "btuel", "cooltype"),
                   force = c("moneypy", "householder_race", "education", "nhsldmem", "kownrent", "recs_division"),
                   note = "Hello world. Reminder: running in test mode by default.")
list.files(dir)

# Using default settings: test mode and no upload to Google Drive
out <- fusionOutput(input = dir)
list.files(out)